Refactor entity packages to improve modularity
move each entity to its responsible module instead of storing all entities in one module for better modularity.
