+++ core/src/main/java/com/rss/backend/account/entity/Role.java
@@ -1,4 +1,4 @@
-package com.rss.backend.domain;
+package com.rss.backend.account.entity;
 
 public enum Role {
     RIDER,
