+++ core/src/main/java/com/rss/backend/account/repository/DriverRepository.java
@@ -1,6 +1,6 @@
 package com.rss.backend.account.repository;
 
-import com.rss.backend.domain.entity.Driver;
+import com.rss.backend.account.entity.Driver;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DriverRepository extends JpaRepository<Driver, Long> {
