+++ core/src/main/java/com/rss/backend/account/repository/RiderRepository.java
@@ -1,6 +1,6 @@
 package com.rss.backend.account.repository;
 
-import com.rss.backend.domain.entity.Rider;
+import com.rss.backend.account.entity.Rider;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RiderRepository extends JpaRepository<Rider, Long> {
