+++ core/src/main/java/com/rss/backend/account/service/AuthService.java
@@ -4,13 +4,13 @@ import com.rss.backend.account.dto.AuthRequest;
 import com.rss.backend.account.dto.AuthResponse;
 import com.rss.backend.account.dto.RegisterDriverRequest;
 import com.rss.backend.account.dto.RegisterRiderRequest;
-import com.rss.backend.domain.*;
+import com.rss.backend.account.entity.Role;
 import com.rss.backend.account.repository.DriverRepository;
 import com.rss.backend.account.repository.RiderRepository;
 import com.rss.backend.account.repository.UserRepository;
-import com.rss.backend.domain.entity.Driver;
-import com.rss.backend.domain.entity.Rider;
-import com.rss.backend.domain.entity.User;
+import com.rss.backend.account.entity.Driver;
+import com.rss.backend.account.entity.Rider;
+import com.rss.backend.account.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
