+++ core/src/main/java/com/rss/backend/map/repository/EdgeRepository.java
@@ -1,6 +1,6 @@
 package com.rss.backend.map.repository;
 
-import com.rss.backend.domain.entity.Edge;
+import com.rss.backend.map.entity.Edge;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
