+++ core/src/main/java/com/rss/backend/map/repository/NodeRepository.java
@@ -1,6 +1,6 @@
 package com.rss.backend.map.repository;
 
-import com.rss.backend.domain.entity.Node;
+import com.rss.backend.map.entity.Node;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface NodeRepository extends JpaRepository<Node, Long> {
